DOCS-1868: Reconfigures OSS release notes, moves component version to ref
