+++ calico/reference/component-versions.mdx
@@ -1,0 +1,8 @@
+---
+description: A list of component versions for Calico
+---
+
+# Component versions
+
+
+<ComponentVersions />
