+++ calico/release-notes/index.mdx
@@ -1,7 +1,68 @@
 ---
-description: What's new, and why features provide value for upgrading.
+description: Release notes for Calico Open Source
 ---
 
-# Release notes
+# Calico Open Source 3.27 release notes
 
-<ReleaseNotes />
+Learn about the new features, bug fixes, and other updates in this release of {{prodname}}.
+
+## New features and enhancements
+
+### HPC-based Windows install and upgrade
+
+We've added support for HPC (HostProcess/Privileged containers) install and upgrade for Windows nodes/
+{{prodname}} can now bootstrap Windows worker nodes just like Linux worker nodes.
+Platform engineers can scale Windows worker nodes with an auto-scaling node pool without bootstraping each node manually, and upgrade {{prodname}} deployments without manually upgrading nodes one at a time.
+
+For more information, see [Quickstart](../getting-started/kubernetes/windows-calico/quickstart#install-calico-for-windows-using-hostprocess-containers).
+
+### ipV6 support for eBPF
+
+eBPF now supports ipv6, which offers advanced and better features compared with ipv4.
+
+For eBPF details, see [Enable the eBPF dataplane](../operations/ebpf/enabling-ebpf.mdx).
+
+### Support for Azure CNI with overlay networking for AKS
+
+We've add supported for Azure CNI with overlay networking for AKS clusters.
+This option augments support for the Azure CNI with no overlay where a VNET IP address is assigned to every pod, which can lead to IP address exhaustion.
+
+To review Azure CNI options, see [Microsoft Azure Kubernetes Service (AKS)](../getting-started/kubernetes/managed-public-cloud/aks).
+
+## Deprecated and removed features
+
+The following features are being deprecated or removed in this release:
+
+## Bug fixes
+
+* Bug 1.
+* Bug 2.
+
+## Security fixes
+
+## Known issues
+
+## Release details
+
+### Calico Open Source 3.27.0 general availability release
+
+XX October 2023
+
+Calico Enterprise 3.18.0 is now available as an early preview release.
+This release is for previewing and testing purposes only.
+It is not supported for use in production.
+
+<!--
+### Calico Open Source 3.18.1 bux fix update
+
+DD M YYYY
+
+#### Enhancements
+
+#### Bug fixes
+
+#### How to update
+
+To update an existing installation of Calico Enterprise 3.18, see .
+
+-->
